Moving models to the models folder in the data folder
