+++ app/src/main/java/ru/androidacademy/gooseeye/adapters/MovieRecyclerAdapter.kt
@@ -5,7 +5,7 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import ru.androidacademy.gooseeye.R
-import ru.androidacademy.gooseeye.data.Movie
+import ru.androidacademy.gooseeye.data.models.Movie
 import ru.androidacademy.gooseeye.databinding.ItemMovieBinding
 
 class MovieRecyclerAdapter(private val listener: (Movie) -> Unit) :
