+++ app/src/main/java/ru/androidacademy/gooseeye/api/responses/GenresResponse.kt
@@ -3,7 +3,7 @@ package ru.androidacademy.gooseeye.api.responses
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import ru.androidacademy.gooseeye.data.Genre
+import ru.androidacademy.gooseeye.data.models.Genre
 
 @ExperimentalSerializationApi
 @Serializable
