+++ app/src/main/java/ru/androidacademy/gooseeye/data/models/Genre.kt
@@ -1,4 +1,4 @@
-package ru.androidacademy.gooseeye.data
+package ru.androidacademy.gooseeye.data.models
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
