+++ app/src/main/java/ru/androidacademy/gooseeye/viewmodels/MovieDetailsViewModel.kt
@@ -4,8 +4,8 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.launch
-import ru.androidacademy.gooseeye.data.Actor
-import ru.androidacademy.gooseeye.data.Movie
+import ru.androidacademy.gooseeye.data.models.Actor
+import ru.androidacademy.gooseeye.data.models.Movie
 import ru.androidacademy.gooseeye.data.Repository
 
 class MovieDetailsViewModel(
